added FireBase Messaging

+++ app/src/main/java/com/example/userproductscart/model/Product.java
@@ -9,7 +9,7 @@ public class Product {
 
     //common
     public String name;
-    public byte type;
+    public int type;
 
     //WeightBased
     public int pricePerKg;
